Preliminary version of the showcase script
This script still contains elements of bad style (e.g. a function defined in the main program) and numerous elements which could be improved (e.g. names and paths, using the rasterio library for exporting rasters, ...). Nevertheless it is provided as a technical demonstration of how necessary operations for a damage assessment can be implemented. The essential steps are: (1) reading input data (2) polygonize (3) intersect and dissolve (4) apply classifiers (5) write results as GeoJSON