Transformation to 4326
Introduced a simple check whether the CRS of all input rasters is identical. Polygonization is still done in the CRS of the rasters, but the final results are transformed to EPSG:4326 before exporting. It is easier to implement this transformation on geopandas objects than to transform the rasters in GDAL, since this would require intermediate files to be stored on disk or in-memory VRT